Debugger rework to support beat and improve speed
Originally, values were retrieved from the 'Listener' by the UI to display, now instead the Listener will add to specialised threads new data from the analysis, the threads are then used to update the UI.
This is also more efficient as the threads will now sleep when there is no data to be added.
Still need to reimplement the Rewind functionality properly and clean up the rest of this code.