Initial work on a profiling script for our library/benchmarking tool.
Originally I started to write together some scripts to benchmark each worker individually.

To me this seemed a bit pointless, and we'd have to stitch together the timings of multiple workers to get an idea of how long it actually takes to get to the end point, i.e. a signal being produced.

This instead times how long it took for a dummy signal to go through the entire hierarchy.

This means users could potentially run the script, using the config options they want to get an idea of the average delay on their machine when running the library how they want to. Allowing them to tweak settings until they've found a configuration with a delay they are happy with :)